+++ vblog/src/main/java/com/hygge/vblog/mapper/VArticleMapper.java
@@ -21,9 +21,9 @@ public interface VArticleMapper extends BaseMapper<VArticle> {
      * @param pageSize
      * @return
      */
-    List<BlogsDto> selectAll(@Param("current")Integer current, @Param("pageSize")Integer pageSize);
+    List<BlogsDto> selectAll(@Param("current")Long current, @Param("pageSize")Long pageSize);
 
-    List<BlogsDto> selectAllFlag(@Param("current")Integer current, @Param("pageSize")Integer pageSize, @Param("flag")Integer flag);
+    List<BlogsDto> selectAllFlag(@Param("current")Long current, @Param("pageSize")Long pageSize, @Param("flag")Integer flag);
 
     /**
      * 保存文章并返回主键id
